requirements.txt
pytorch-lightning==1.5.6
PyYAML==6.0
tensorboard>=2.2.0
torch>=1.6.1
torchsummary==1.5.1
torchvision>=0.10.1

gdown
pytest